Перенос логики работы со сценами в SceneLoader Перенос функционала, отвечающего за гонку в RaceManager. Связывание RaceManager в LocationInstaller Переименовывание во всех классах gameManager в raceManager
