+++ Assets/Scripts/Game/RaceManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f18df09c2f3003e428bb52ab5dc7a1e5
+guid: 7d0ae23b0e18caa4eac90383a225a906
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
