+++ Assets/Scripts/Service/PauseService.cs
@@ -2,14 +2,20 @@ using UnityEngine;
 
 public class PauseService
 {
-    private bool _isPaused = false;
-    private float _prevTimeScale = 1f;
+    private bool _isPaused;
+    private float _normalTimeScale;
+
+    public PauseService()
+    {
+        _isPaused = false;
+        _normalTimeScale = 1f;
+    }
+
     public void SetPause()
     {
         if (!_isPaused)
         {
             _isPaused = true;
-            _prevTimeScale = Time.timeScale;
             Time.timeScale = 0f;
         }
     }
@@ -18,7 +24,7 @@ public class PauseService
         if (_isPaused)
         {
             _isPaused = false;
-            Time.timeScale = _prevTimeScale;
+            Time.timeScale = _normalTimeScale;
         }
     }
 }
